ADD: main component Keyword List component 분리
Modify: Conflice Resolve
